+++ .env
@@ -1,2 +1,3 @@
-HOST=apptware-docs-production.up.railway.app
+# HOST=apptware-docs-production.up.railway.app
+HOST=localhost:3000
 PORT=3000
