+++ package.json
@@ -3,14 +3,20 @@
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
+  "scripts": {
+    "start": "node index.js",
+    "dev": "nodemon index.js"
+  },
   "dependencies": {
     "angular-expressions": "^1.4.3",
+    "cors": "^2.8.5",
     "docxtemplater": "^3.58.1",
     "dotenv": "^16.4.7",
     "express": "^4.17.1",
     "lodash": "^4.17.21",
-    "nodemon": "^3.1.9",
-    "pizzip": "^3.1.7",
-    "cors": "^2.8.5"
+    "pizzip": "^3.1.7"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.9"
   }
 }
